+++ fastlane/lib/fastlane/actions/setup_travis.rb
@@ -1,17 +1,15 @@
-require 'fastlane/ci'
-
 module Fastlane
   module Actions
     class SetupTravisAction < Action
       def self.run(params)
         # Stop if not executed by CI
-        unless Ci.should_run?(force: params[:force])
+        unless Helper::CI.should_run?(force: params[:force])
           UI.message("Currently not running on CI system, skipping travis setup")
           return
         end
 
         # Create a temporary keychain
-        Ci.setup_keychain
+        Helper::CI.setup_keychain
       end
 
       #####################################################
